Support App::cpm for build script

+++ build_all_cpanm.sh
@@ -3,7 +3,9 @@
 cpanm --prompt Module::Install
 for module_name in Whatbot Whatbot-*/ ; do
     cd $module_name
-    rm Makefile
-    cpanm --prompt .
+    if [ -e Makefile ]; then
+    	rm Makefile
+    fi
+    cpanm -n --installdeps .
     cd ..
 done
